small indexer refactoring
